requirements.txt
langchain
langchain-openai
langchain-google
pydantic
python-dotenv
langgraph
wikipedia
langchain-community
langchain_google_genai
langgraph-checkpoint-sqlite
pypdf
langchain-postgres
langchain_redis